Use -ko-attr-alt-if instead of ternary operator for removing alt attribute when empty
This allow mosaico to read the default value for the variable and resolve a "minor" model incompatibility introduced with versafix-template-1.1.18 (while fixing #9).

This reintroduces compatibility with 1.1.16.

Also made imageHeight to extend height like it was in previous releases (useless but the compatibility check code is not so advanced to understand this)